Implement: Define party member level/job change event types [#62]
- Updated character/kafka.go to import job package
- Changed memberJobChangedEventBody to use job.Id instead of uint16
- Updated jobChangedEventProvider function signature to use job.Id
- Updated processor.go to remove unnecessary type conversions
- Ensured proper type consistency across the event system

+++ atlas.com/parties/character/processor.go
@@ -179,7 +179,7 @@ func (p *ProcessorImpl) JobChange(mb *message.Buffer) func(characterId uint32, j
 		
 		// If character is in a party, emit party member job changed event
 		if c.PartyId() != 0 {
-			err = mb.Put(EnvEventMemberStatusTopic, jobChangedEventProvider(c.PartyId(), c.WorldId(), characterId, uint16(oldJobId), uint16(jobId), c.Name()))
+			err = mb.Put(EnvEventMemberStatusTopic, jobChangedEventProvider(c.PartyId(), c.WorldId(), characterId, oldJobId, jobId, c.Name()))
 			if err != nil {
 				p.l.WithError(err).Errorf("Unable to announce the party [%d] member [%d] job changed.", c.PartyId(), c.Id())
 				return err
